WIP Functions which change obj props moved to class methods
Any functions which pertained to changing instantiated objects are now
class methods. This makes the code (relatively) easier to follow. The
goal is to get everything that directly changes objects into the classes
and use etiquette.py to handle just the logic.

Added extensive comments to etiquette.py, as it still is not clear from
the code what is being done. (Fix this before merge to master.)