etiquette: Add initialise() and run_cycles(); replaces globals
initialise() instantiates the UrinalSession, StallRow, and Stall classes
(or list of objects, in the case of Stall). The instantiation of these
classes had previously created global variables. As this app no longer
uses global variables, it is safer to do it this way.

run_cycles() contains all of the logic which ran the app. This too was
done globally before.